+++ server/src-lib/Hasura/RQL/Types.hs
@@ -58,8 +58,11 @@ import Hasura.RQL.Types.Network as R
 import Hasura.RQL.Types.Permission as R
 import Hasura.RQL.Types.QueryCollection as R
 import Hasura.RQL.Types.QueryTags as R
-import Hasura.RQL.Types.Relationship as R
-import Hasura.RQL.Types.RemoteRelationship as R
+import Hasura.RQL.Types.Relationships.FromSource as R
+import Hasura.RQL.Types.Relationships.Local as R
+import Hasura.RQL.Types.Relationships.Remote as R
+import Hasura.RQL.Types.Relationships.ToSchema as R
+import Hasura.RQL.Types.Relationships.ToSource as R
 import Hasura.RQL.Types.RemoteSchema as R
 import Hasura.RQL.Types.ResultCustomization as R
 import Hasura.RQL.Types.Roles as R
@@ -372,7 +375,7 @@ askRelType m r msg = do
 askRemoteRel ::
   (MonadError QErr m) =>
   FieldInfoMap (FieldInfo backend) ->
-  RemoteRelationshipName ->
+  RelName ->
   m (RemoteFieldInfo backend)
 askRemoteRel fieldInfoMap relName = do
   fieldInfo <- askFieldInfo fieldInfoMap (fromRemoteRelationship relName)